Implement surface normal shading
Change `hit_sphere` to return the actual `t` at which the ray hits the
sphere, so that we can compute the surface normal based on the hit point
and the center of the sphere.

Change `ray_colour` to compute the surface normal and map its components
to a colour, in order to shade it in on the rendered image.