Integrate `Hittable` and a world into `main`
The world is just a `Hittable_List` into which we cast a ray, then move
the ray a bit, then cast again. Hence why it's good to have a list of
`Hittable`!

(And we can "cheat" and create the ground by adding a large sphere
that the camera can only see the top of ^^)